Refactor: Impl 구조로 변경 (#14)
